feature: добавить модули
