+++ js/utils/data.js
@@ -1,0 +1,41 @@
+const TYPES= [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+  'hotel',
+];
+
+const TIME = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+
+const FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+
+const PHOTOS = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
+];
+
+const DESCRIPTIONS = [
+  'красивое светлое и мухи не кусают. Ну и конечно, евроремонт',
+  'для славян без кошек и детей - Кекс не рекомендует',
+  'можно всех и с детьми, и с кошкамии, и на девичник тоже',
+];
+
+export{TYPES};
+export{TIME};
+export{FEATURES};
+export{PHOTOS};
+export{DESCRIPTIONS};
+
